+++ src/NCalc/Exceptions/NCalcEvaluationException.cs
@@ -11,7 +11,17 @@ public class NCalcEvaluationException : NCalcException
         Location = ExpressionLocation.Empty;
     }
 
+    public NCalcEvaluationException(string message, Exception innerException) : base(message, innerException)
+    {
+        Location = ExpressionLocation.Empty;
+    }
+
     public NCalcEvaluationException(string message, ExpressionLocation location) : base(message)
+    {
+        Location = location;
+    }
+
+    public NCalcEvaluationException(string message, ExpressionLocation location, Exception innerException) : base(message, innerException)
     {
         Location = location;
     }